+++ src/components/article/layout/title-row-upon.js
@@ -1,4 +1,4 @@
-import Link from 'react-router/lib/Link'
+import Link from 'react-router-dom/Link'
 import PropTypes from 'prop-types'
 import React from 'react'
 import constPageThemes  from '../../../constants/page-themes'
@@ -57,7 +57,7 @@ const HeaderElementBlock = styled.div`
   display: block;
 `
 
-const Title = HeaderElementBlock.extend`
+const Title = styled(HeaderElementBlock)`
   margin: 0;
   font-weight: ${typography.font.weight.bold};
   line-height: 1.4;
@@ -68,7 +68,7 @@ const Title = HeaderElementBlock.extend`
   `}
 `
 
-const Subtitle = HeaderElementBlock.extend`
+const Subtitle = styled(HeaderElementBlock)`
   color: ${props => (colorselector(props, colors.gray.gray50))};
   font-size: ${typography.font.size.medium};
   font-weight: ${typography.font.weight.extraLight};
@@ -80,7 +80,7 @@ const Subtitle = HeaderElementBlock.extend`
   `}
 `
 
-const Topic = HeaderElementBlock.extend`
+const Topic = styled(HeaderElementBlock)`
   > a {
     text-decoration: none !important;
     border: 0 !important;