+++ src/constants/category.js
@@ -1,0 +1,42 @@
+const pathSegments = {
+  cultureAndArt: 'culture_and_art',
+  environmentAndEducation: 'environment_and_education',
+  humanRightsAndSociety: 'human_rights_and_society',
+  infographic: 'infographic',
+  international: 'international',
+  livingAndMedicalCare: 'living_and_medical_care',
+  photography: 'photography',
+  politicsAndEconomy: 'politics_and_economy',
+  reviews: 'reviews'
+}
+
+export default {
+  ids: {
+    [pathSegments.cultureAndArt]: '57175d923970a5e46ff854db',
+    [pathSegments.environmentAndEducation]: '5951db9b507c6a0d00ab063d',
+    [pathSegments.humanRightsAndSociety]: '5951db87507c6a0d00ab063c',
+    [pathSegments.infographic]: '59671cd2507c6a0d00ab06ba',
+    [pathSegments.international]: '5743d35a940ee41000e81f4a',
+    [pathSegments.livingAndMedicalCare]: '59783ad89092de0d00b41691',
+    [pathSegments.photography]: '574d028748fa171000c45d48',
+    [pathSegments.politicsAndEconomy]: '5951dbc2507c6a0d00ab0640',
+    [pathSegments.reviews]: '573177cb8c0c261000b3f6d2'
+  },
+  labels: {
+    cultureAndArt: '文化．藝術',
+    [pathSegments.cultureAndArt]: '文化．藝術',
+    environmentAndEducation: '環境．教育',
+    [pathSegments.environmentAndEducation]: '環境．教育',
+    humanRightsAndSociety: '人權．社會',
+    [pathSegments.humanRightsAndSociety]: '人權．社會',
+    [pathSegments.infographic]: '多媒體',
+    [pathSegments.international]: '國際．兩岸',
+    livingAndMedicalCare: '生活．醫療',
+    [pathSegments.livingAndMedicalCare]: '生活．醫療',
+    politicsAndEconomy: '政經．產業',
+    [pathSegments.politicsAndEconomy]: '政經．產業',
+    [pathSegments.reviews]: '評論',
+    [pathSegments.photography]: '攝影'
+  },
+  pathSegments
+}