+++ src/constants/redux-state-prop-key.js
@@ -10,7 +10,6 @@ export default {
   selectedPost: reduxStateFields.selectedPost,
   selectedTopic: reduxStateFields.selectedTopic,
   bookmarks: 'bookmarks',
-  routing: 'routing',
   header: 'header',
   searchedAuthorsList: 'searchedAuthorsList',
   authorsList: 'authorsList',