+++ src/data-loaders/index.js
@@ -1,0 +1,21 @@
+import loadArticlePageData from './article-page'
+import loadAuthorListPageData from './author-list-page'
+import loadAuthorPageData from './author-page'
+import loadCategoryListPageData from './category-list-page'
+import loadIndexPageData from './index-page'
+import loadPhotographyPageData from './photography-page'
+import loadTagListPageData from './tag-list-page'
+import loadTopicListPageData from './topic-list-page'
+import loadTopicPageData from './topic-page'
+
+export default {
+  loadArticlePageData,
+  loadAuthorListPageData,
+  loadAuthorPageData,
+  loadCategoryListPageData,
+  loadIndexPageData,
+  loadPhotographyPageData,
+  loadTagListPageData,
+  loadTopicListPageData,
+  loadTopicPageData
+}